+++ src/drv/cdpdf.c
@@ -37,7 +37,7 @@ struct _cdCtxCanvas
   double height_mm;      /* Altura do papel (mm) */  
   double scale;          /* Fator de conversao de coordenadas (pixel2points) */
   int landscape;         /* page orientation */
-  float  rotate_angle;
+  double rotate_angle;
   int    rotate_center_x,
          rotate_center_y;
 
@@ -156,7 +156,7 @@ static void sUpdateFill(cdCtxCanvas *ctxcanvas, int fill)
                                                   get_blue(ctxcanvas->canvas->foreground), 0);
     }
     else
-      PDF_setcolor(ctxcanvas->pdf, "fill", "pattern", (float)ctxcanvas->pattern, 0, 0, 0);
+      PDF_setcolor(ctxcanvas->pdf, "fill", "pattern", (double)ctxcanvas->pattern, 0, 0, 0);
   }
 }
 
@@ -1162,7 +1162,7 @@ static void cdtransform(cdCtxCanvas *ctxcanvas, const double* matrix)
   {
     /* rotation = translate to point + rotation + translate back */
     PDF_translate(ctxcanvas->pdf, ctxcanvas->rotate_center_x, ctxcanvas->rotate_center_y);
-    PDF_rotate(ctxcanvas->pdf, (double)ctxcanvas->rotate_angle);
+    PDF_rotate(ctxcanvas->pdf, ctxcanvas->rotate_angle);
     PDF_translate(ctxcanvas->pdf, -ctxcanvas->rotate_center_x, -ctxcanvas->rotate_center_y);
   }
 }
@@ -1386,9 +1386,9 @@ static void set_rotate_attrib(cdCtxCanvas *ctxcanvas, char* data)
 
   if (data)
   {
-    sscanf(data, "%g %d %d", &ctxcanvas->rotate_angle,
-                             &ctxcanvas->rotate_center_x,
-                             &ctxcanvas->rotate_center_y);
+    sscanf(data, "%lg %d %d", &ctxcanvas->rotate_angle,
+                              &ctxcanvas->rotate_center_x,
+                              &ctxcanvas->rotate_center_y);
   }
   else
   {
@@ -1407,7 +1407,7 @@ static char* get_rotate_attrib(cdCtxCanvas *ctxcanvas)
   if (!ctxcanvas->rotate_angle)
     return NULL;
 
-  sprintf(data, "%g %d %d", (double)ctxcanvas->rotate_angle,
+  sprintf(data, "%g %d %d", ctxcanvas->rotate_angle,
                             ctxcanvas->rotate_center_x,
                             ctxcanvas->rotate_center_y);
 
@@ -1631,7 +1631,7 @@ static void cdcreatecanvas(cdCanvas* canvas, void *data)
 
     if (*line != '\0')
     {
-      float num;
+      double num;
       line++;
       switch (*line++)
       {
@@ -1645,12 +1645,12 @@ static void cdcreatecanvas(cdCanvas* canvas, void *data)
           break;
         }
       case 'w':
-        sscanf(line, "%g", &num);
+        sscanf(line, "%lg", &num);
         ctxcanvas->width_mm = num;
         ctxcanvas->width_pt = CD_MM2PT*ctxcanvas->width_mm;
         break;
       case 'h':
-        sscanf(line, "%g", &num);
+        sscanf(line, "%lg", &num);
         ctxcanvas->height_mm = num;
         ctxcanvas->height_pt = CD_MM2PT*ctxcanvas->height_mm;
         break;
